Created API Folder For Requests, Refactored Action
Created APIs folder to hold external files that will make API requests
which will then be imported into our action and called from there with
the appropriate endpoints (actions then get called in componentDidMount
methods)
- Import axios
- Export axios create with an object containing k/v pair for the baseURL

Refactored action creator for fetchPosts to be async, since it is now
making an API call
- Import API file
- Refactor fetchPosts to be async, i.e., make function async, refactor
  API call to use await on the jsonPlaceholder.get(endpoint) method
  and assign this to a response variable
- Assign response to payload property of returned action

THIS IS A BAD APPROACH, THIS COMMIT IS TO ILLUSTRATE WHY!!!

Because of the way that Babel transpiles React / ES6+ code to code that
can be interpreted by legacy browsers, we run into an error. Actions
must return POJO (plain old javascript object) with a type property.
When our async/await code gets transpiled we enter a set of switch
statements, and our return value is not the object with the type
property but rather our jsonPlaceholder.get() function. Our code never
even sees the case with the object with the type property. Additionally
by the time that this action gets dispatched and reaches our reducer,
we have not retrieved data from our API. Refactoring into a promise
will not work either. This request happens so fast that we can't get a
response from the server in the time that the dispatch happens. In order
to resolve this, we need to make use of redux thunk