Refactored Actions To Make Use Of Thunk
As demonstrated in the previous commit, without middlewares, we cannot
make use of asynchronous action creators because action creators must
return Objects and not functions, and because at 'compile' time, or
rather, when the API request gets made, it happens so fast that
we are not able to asynchronously handle the response. In order to
resolve this, we are make use of the middleware redux thunk. Redux thunk
is a very simple middleware and all it does is check to see whether our
action returns a function or an object. If our action returns a function
redux thunk takes the function, calls it with dispatch and getState
passed as arguments, allowing us to dispatch the function at our leisure
i.e., once the API response has been resolved. Once the API response has
been resolved, we are going to dispatch the action manually, get a new
action, and then enter the the redux thunk loop again. Once the new
action has been dispatched, it returns an object, and so it gets passed
on to the reducers.

First, we wire up redux thunk in our root index.js
- Import thunk from redux-thunk
- Import applyMiddlewares from react-redux
- Pull out store from Provider's props into its own variable, pass store
  variable into store props i.e. store={store}
  - We do this in order to pass applyMiddleware called with thunk
    passed as a second argument to createStore

Then, we refactor our action creator
  - Refactor fetchPosts to be an anonymous arrow function that returns
    an async function
    - We are able to use async for this inner function! We can use
      async await because redux thunk is handling the return value of
      a function
    - Return function gets called with dispatch and getState
    - Return function returns an await response assigned to a function
      call for our API request and dispatch called with the action
      type associated with the action and a payload of the response
      body
  - Refactor this code to align with ES6
    - Remove getState since we are not using it
    - One argument for an arrow function so remove parens around it
    - Remove return and curly braces around inner function

No errors in browser, refactor successful.