Installed Lodash, Memoized fetchUser
In order to solve overfetching issue (i.e., fetching user from the
UserHeader component each time a post is rendered to the DOM) we can
memoize our fetchUser function. Memoizing is essentially creating a
memory cache of the request, so that when the request is made to
an endpoint like /users/1, /users/2, etc that request fn is put into a
cache and can be called again but when it is called again it no longer
makes the request.

The extremely popular lodash library has a built in memoization function
which we are going to implement in our App for memoizing fetchUser
requests. We have installed the library.
- By convention, lodash is imported into apps as the character _ and

We attempted refactoring our fetchUser function by expanding our arrow
functions to expanded / ES5 syntax functions and implemented the lodash
memoize function called on the entire fetchUser function. This still
makes 100 fetchUser requests
- In order to properly refactor our code, we make a memoized fetchUser
  function, named _fetchUser to denote its status as a private function
  - _fetchUser is defined as _.memoize() with the inner function of
    fetchUser passed as a callback
  - The inner function gets passed id and dispatch
  - Refactor fetchUser to return the memoized _fetchUser called with
    id and dispatch
- Our app now only makes one request to the API for each user ID